Update docs and re-organize tests
